requirement.txt
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.5
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-tinymce4-lite==1.7.5
idna==2.9
jsmin==2.2.2
oauthlib==3.1.0
Pillow==7.1.1
pkg-resources==0.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.8
